+++ R/modelling.R
@@ -324,7 +324,7 @@ bestModel <- function(data,
     message("\nThe model includes a high number of terms. Simpler models are usually more robust. Cross validation with 'cv(model$data)' or an inspection of information functions with 'plot.subset' might help to identify a balanced number of terms. Consider fixing this parameter to a smaller number.")
   }
 
-  message("\nUse 'print(model)' to get detailed information on the different solutions, 'plot(model) to display percentile plot, plot.subset(model)' to inspect model fit.")
+  message("\nUse 'printSubset(model)' to get detailed information on the different solutions, 'plotPercentiles(model) to display percentile plot, plotSubset(model)' to inspect model fit.")
 
   if(plot&&attr(data, "useAge")&&(attributes(data)$group %in% colnames(data))){
     plotPercentiles(data, bestformula)
@@ -364,6 +364,7 @@ printSubset <- function(x, ...) {
       BIC = x$subsets$bic,
       Terms = x$subsets$numberOfTerms
     ))
+  invisible(x)
   return(table)
 }